refactor: strategy job specific resource
