+++ simaple/data/jobs/definitions/__init__.py
@@ -1,3 +1,4 @@
+from simaple.data.jobs.definitions.builtin_strategy import BuiltinStrategy
 from simaple.data.jobs.definitions.passive_hyper_skill import (
     MultiplierPassiveHyperskill,
     PassiveHyperskillInterface,
@@ -12,4 +13,5 @@ __all__ = [
     "MultiplierPassiveHyperskill",
     "StatIncreasePassiveHyperskill",
     "ValueIncreasePassiveHyperskill",
+    "BuiltinStrategy",
 ]
