+++ simaple/data/jobs/resources/builtin_strategy/adele.yaml
