+++ simaple/data/jobs/resources/builtin_strategy/archmagefb.yaml
