+++ simaple/data/jobs/resources/builtin_strategy/archmagetc.yaml
