+++ simaple/data/jobs/resources/builtin_strategy/bishop.yaml
