+++ simaple/data/jobs/resources/builtin_strategy/dualblade.yaml
