+++ simaple/data/jobs/resources/builtin_strategy/mechanic.yaml
