+++ simaple/data/jobs/resources/builtin_strategy/soulmaster.yaml
