+++ simaple/data/jobs/resources/builtin_strategy/windbreaker.yaml
